wlan: Disable NS offload for used slot index instead of all slots
Currently, Host driver disable NS offload for all the slots index
instead of actually used slots. Because of this host is sending
many unnecessary commands to firmware. Hence firmware is
entering and exit BMPS many time internally which is leading
to interrupts lock in firmware, which lead to delay in scheduling
of BT thread. Eventually user is getting A2DP glitches in BT.

As a part of this fix, Host stores used slots and disable NS offload
For these slots only.

Change-Id: Iaebf0971758fea116c398f1034028b0537eca119
CRs-Fixed: 2069206