dwc3: Do not perform core reinitialisation for host only mode
Cuurently as a part of start host, after the block reset we
are performing core init and gadget restart of the dwc3
controller. This causes crash for host only controller where
ep0 and ep1 have not been initialised but are being
dereferenced. Also, HW reinitialisation is not required for
start or restart host routine.

Change-Id: Ie2b781ecb03a1d0a02c5f4305d518adc18dc6ade
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>